InfinityInvest-main/requirements.txt
random-password-generator==2.1.0
django==2.2.7
djangorestframework==3.10.2
requests==2.20.0
validate-docbr==1.8.2